qNoise: A Generator of non-Gaussian Colored Noise
This is a non-gaussian colored noise depending on only two parameters.
Tau which controls the autocorrelation of the noise (tau=0 recovers the
white noise limit) and q is a parameter for the statistics of the noise
(q<1 is bounded domain noise, q > is supra-Gaussian noise and q=1
recovers the gaussian noise limit). It is very useful for the use on
simulations or as a source of controlled noise for experiments.